refactor: points 일글컬렉션 추출
